requirements.txt
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-cors-headers
django-environ
django-smtp-ssl
django-storages
djangorestframework
djangorestframework-simplejwt
django-mptt
django-tinymce
gunicorn
idna
jmespath
oauth2-provider
oauthlib
packaging
Pillow
psycopg2-binary
pycparser
PyJWT
python-dateutil
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
s3transfer
six
sqlparse
tzdata
urllib3
whitenoise
drf-yasg
djoser
django-ses
django-ses[events]
social-auth-app-django
facebook-business
stripe